+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support inline eval of macros with `#` and `##` syntax
 - Support onhover/getDefinition/getDecleration on Macro param
 
+### Changed
+
+- Remove `required` property `device_type` for `memory` and `cpu` node
+
 # 0.2.1 - 2025-03-23
 
 ### Added