+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+# 0.2.3 - 2025-04-02
+
+### Fixed
+
+- Issue with folding ranges not showing in all context files
+
 # 0.2.2 - 2025-04-01
 
 ### Added